fix regression that introduced race condition build failures for gcc
commit 20ce424e65ab227e3dc866ab6ac2d2b096f55fd8 introduced a situation
where two instances of gcc's build process could be invoked
concurrently in the process of fixing the same issue for musl. only
the production of native compilers seems to have been affected. add a
direct dependency edge rather than relying on the indirect dependency,
which was only present in non-native mode.