make --disable-separate-code the default for binutils
the default is ordered such that user-provided config variables in
config.mak or on the make command line can still override it.

this is a dubious anti-ROP feature with high cost (file size, load
time, VMA count consumed per library), and historically was broken in
some binutils versions. the ones we use don't seem to be affected, but
it's better to have it off anyway.