Add license file and copyright notes
After discussion with Rich Felker and other contributors, it's clear
that we want to license this project under the MIT/Expat license.
However, we need to be clear that we're not claiming to license the
binary artifacts, as those retain the licenses of the upstream
projects.  Similarly, the patches, to the extent that there is any
copyright interest in them at all, retain the licenses from the
upstream projects.

We've added language in the README and in a new `COPYRIGHT` file to
make all of this clear.

Ref: #25 #30 #80