Fix initialization of CoW source trees
As part of creating a CoW source tree, we symlink all top-level paths
from the original directory.  In commit ebf93035c, we removed a shell
glob used for this and replaced it with an invocation of `find`.

However, our invocation of `find ... -exec sh -c '... $@ ...' {} \+`
is causing the first path in every directory to be skipped, breaking
the build.  This is because arguments to `sh -c ...` begin with the
zeroth argument, while `$@` only returns the first argument onward.

Let's fix this by providing an explicit zeroth argument to `sh -c`.