fix wrong ac_cv_prog_lex_root value
the value should only be the root, not including the .c suffix. as
noted in commit 2d0015f547b1846eeab1359f5e8ce493d2be50f9 which added
this, the result is not actually used, so it doesn't matter that it's
wrong, but the wrong value produces spurious error output in configure
logs.