fix cowpatch issue with libstdc++-v3 directory in gcc
this directory uses relative pathnames to access libgcc unwind
headers, but with cowpatch, those are relative to the libstdc++-v3
directory in the unpatched tree unless libstdc++-v3 has been
cow-split.

this calls for a general solution at some point, but for now, just add
an explicit dummy patch to trigger cow splitting.